+++ src/DurableTask.SqlServer/Scripts/logic.sql
@@ -1,6 +1,85 @@
 ﻿-- Copyright (c) Microsoft Corporation.
 -- Licensed under the MIT License.
 
+-- Create custom types. This must be done before creating stored procedures.
+-- IMPORTANT: If you make any changes to these types, you must also add a line in the schema upgrade script
+--            (for example, schema-1.X.0) to drop them so that those changes are properly reflected.
+IF TYPE_ID(N'__SchemaNamePlaceholder__.InstanceIDs') IS NULL
+    CREATE TYPE __SchemaNamePlaceholder__.InstanceIDs AS TABLE (
+        [InstanceID] varchar(100) NOT NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.MessageIDs') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.MessageIDs AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [SequenceNumber] bigint NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.HistoryEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.HistoryEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [SequenceNumber] bigint NULL,
+        [EventType] varchar(40) NULL,
+        [Name] varchar(300) NULL,
+        [RuntimeStatus] varchar(30) NULL,
+        [TaskID] int NULL,
+        [Timestamp] datetime2 NULL,
+        [IsPlayed] bit NULL,
+        [VisibleTime] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [ParentInstanceID] varchar(100) NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.OrchestrationEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.OrchestrationEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [SequenceNumber] bigint NULL,
+        [EventType] varchar(40) NULL,
+        [Name] varchar(300) NULL,
+        [RuntimeStatus] varchar(30) NULL,
+        [TaskID] int NULL,
+        [VisibleTime] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [ParentInstanceID] varchar(100) NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+IF TYPE_ID(N'__SchemaNamePlaceholder__.TaskEvents') IS NULL
+    -- WARNING: Reordering fields is a breaking change!
+    CREATE TYPE __SchemaNamePlaceholder__.TaskEvents AS TABLE (
+        [InstanceID] varchar(100) NULL,
+        [ExecutionID] varchar(50) NULL,
+        [Name] varchar(300) NULL,
+        [EventType] varchar(40) NULL,
+        [TaskID] int NULL,
+        [VisibleTime] datetime2 NULL,
+        [LockedBy] varchar(100) NULL,
+        [LockExpiration] datetime2 NULL,
+        [Reason] varchar(max) NULL,
+        [PayloadText] varchar(max) NULL,
+        [PayloadID] uniqueidentifier NULL,
+        [Version] varchar(100) NULL,
+        [TraceContext] varchar(800) NULL
+    )
+GO
+
+
 CREATE OR ALTER FUNCTION __SchemaNamePlaceholder__._CurrentTaskHub(@TaskHubName varchar(150))
     RETURNS varchar(50)
     WITH EXECUTE AS CALLER
@@ -125,6 +204,7 @@ AS
         I.[LastUpdatedTime],
         I.[CompletedTime],
         I.[RuntimeStatus],
+        I.[TraceContext],
         (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = I.[TaskHub] AND
             P.[InstanceID] = I.[InstanceID] AND
@@ -136,7 +216,8 @@ AS
         (SELECT TOP 1 [Text] FROM Payloads P WHERE 
             P.[TaskHub] = I.[TaskHub] AND
             P.[InstanceID] = I.[InstanceID] AND
-            P.[PayloadID] = I.[OutputPayloadID]) AS [OutputText]
+            P.[PayloadID] = I.[OutputPayloadID]) AS [OutputText],
+        I.[ParentInstanceID]
     FROM Instances I
     -- like operator is the simplest way to keep indexed seek with conditional where
     WHERE I.[TaskHub] LIKE __SchemaNamePlaceholder__._CurrentTaskHub('%')
@@ -156,6 +237,7 @@ AS
 	    H.[Name],
 	    H.[RuntimeStatus],
         H.[VisibleTime],
+        H.[TraceContext],
 	    (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = H.[TaskHub] AND
             P.[InstanceID] = H.[InstanceID] AND
@@ -174,6 +256,7 @@ CREATE OR ALTER PROCEDURE __SchemaNamePlaceholder__.CreateInstance
     @InputText varchar(MAX) = NULL,
     @StartTime datetime2 = NULL,
     @DedupeStatuses varchar(MAX) = 'Pending,Running',
+    @TraceContext varchar(800) = NULL,
     @TaskHubName varchar(150) = NULL
 AS
 BEGIN
@@ -241,7 +324,8 @@ BEGIN
         [InstanceID],
         [ExecutionID],
         [RuntimeStatus],
-        [InputPayloadID])
+        [InputPayloadID],
+        [TraceContext])
     VALUES (
         @Name,
         @Version,
@@ -249,7 +333,8 @@ BEGIN
         @InstanceID,
         @ExecutionID,
         @RuntimeStatus,
-        @InputPayloadID
+        @InputPayloadID,
+        @TraceContext
     )
 
     INSERT INTO NewEvents (
@@ -260,6 +345,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [EventType],
+        [TraceContext],
         [PayloadID]
     ) VALUES (
         @Name,
@@ -269,6 +355,7 @@ BEGIN
         @RuntimeStatus,
         @StartTime,
         @EventType,
+        @TraceContext,
         @InputPayloadID)
 
     COMMIT TRANSACTION
@@ -306,7 +393,8 @@ BEGIN
         (CASE WHEN @GetInputsAndOutputs = 0 THEN NULL ELSE P.[Text] END) AS [PayloadText],
         [PayloadID],
         @ParentInstanceID as [ParentInstanceID],
-        @Version as [Version]
+        @Version as [Version],
+        H.[TraceContext]
     FROM History H WITH (INDEX (PK_History))
         LEFT OUTER JOIN Payloads P ON
             P.[TaskHub] = @TaskHub AND
@@ -572,7 +660,6 @@ BEGIN
             E.[InstanceID] = I.[InstanceID]
     WHERE
         I.TaskHub = @TaskHub AND
-        I.[RuntimeStatus] NOT IN ('Suspended') AND
 	    (I.[LockExpiration] IS NULL OR I.[LockExpiration] < @now) AND
         (E.[VisibleTime] IS NULL OR E.[VisibleTime] < @now)
 
@@ -595,9 +682,10 @@ BEGIN
         P.[PayloadID],
         DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime],
         @parentInstanceID as [ParentInstanceID],
-        @version as [Version]
+        @version as [Version],
+        N.[TraceContext]
     FROM NewEvents N
-        LEFT OUTER JOIN [Payloads] P ON 
+        LEFT OUTER JOIN __SchemaNamePlaceholder__.[Payloads] P ON 
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = N.[InstanceID] AND
             P.[PayloadID] = N.[PayloadID]
@@ -634,7 +722,8 @@ BEGIN
         (CASE WHEN [EventType] IN ('TaskScheduled', 'SubOrchestrationInstanceCreated') THEN NULL ELSE P.[Text] END) AS [PayloadText],
         [PayloadID],
         @parentInstanceID as [ParentInstanceID],
-        @version as [Version]
+        @version as [Version],
+        H.[TraceContext]
     FROM History H WITH (INDEX (PK_History))
         LEFT OUTER JOIN Payloads P ON
             P.[TaskHub] = @TaskHub AND
@@ -787,14 +876,16 @@ BEGIN
         [ExecutionID],
         [Name],
         [Version],
-        [RuntimeStatus])
+        [RuntimeStatus],
+        [TraceContext])
     SELECT DISTINCT
         @TaskHub,
         E.[InstanceID],
         NEWID(),
         SUBSTRING(E.[InstanceID], 2, CHARINDEX('@', E.[InstanceID], 2) - 2),
         '',
-        'Pending'
+        'Pending',
+        E.[TraceContext]
     FROM @NewOrchestrationEvents E
     WHERE LEFT(E.[InstanceID], 1) = '@'
         AND CHARINDEX('@', E.[InstanceID], 2) > 0
@@ -802,7 +893,7 @@ BEGIN
             SELECT 1
             FROM Instances I
             WHERE [TaskHub] = @TaskHub AND I.[InstanceID] = E.[InstanceID])
-    GROUP BY E.[InstanceID]
+    GROUP BY E.[InstanceID], E.[TraceContext]
     ORDER BY E.[InstanceID] ASC
 
     -- Create sub-orchestration instances
@@ -813,7 +904,8 @@ BEGIN
         [Name],
         [Version],
         [ParentInstanceID],
-        [RuntimeStatus])
+        [RuntimeStatus],
+        [TraceContext])
     SELECT DISTINCT
         @TaskHub,
         E.[InstanceID],
@@ -821,7 +913,8 @@ BEGIN
         E.[Name],
         E.[Version],
         E.[ParentInstanceID],
-        'Pending'
+        'Pending',
+        E.[TraceContext]
     FROM @NewOrchestrationEvents E
     WHERE E.[EventType] IN ('ExecutionStarted')
         AND NOT EXISTS (
@@ -852,6 +945,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     ) 
     SELECT 
@@ -863,6 +957,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     FROM @NewOrchestrationEvents
     
@@ -891,6 +986,7 @@ BEGIN
         [Name],
         [RuntimeStatus],
         [VisibleTime],
+        [TraceContext],
         [DataPayloadID])
     SELECT
         @TaskHub,
@@ -904,6 +1000,7 @@ BEGIN
         H.[Name],
         H.[RuntimeStatus],
         H.[VisibleTime],
+        H.[TraceContext],
         H.[PayloadID]
     FROM @NewHistoryEvents H
 
@@ -918,7 +1015,8 @@ BEGIN
         [LockedBy],
         [LockExpiration],
         [PayloadID],
-        [Version]
+        [Version],
+        [TraceContext]
     )
     OUTPUT
         INSERTED.[SequenceNumber],
@@ -933,7 +1031,8 @@ BEGIN
         [LockedBy],
         [LockExpiration],
         [PayloadID],
-        [Version]
+        [Version],
+        [TraceContext]
     FROM @NewTaskEvents
 
     COMMIT TRANSACTION
@@ -990,20 +1089,22 @@ BEGIN
             [ExecutionID],
             [Name],
             [Version],
-            [RuntimeStatus])
+            [RuntimeStatus],
+            [TraceContext])
         SELECT DISTINCT
             @TaskHub,
             E.[InstanceID],
             NEWID(),
             SUBSTRING(E.[InstanceID], 2, CHARINDEX('@', E.[InstanceID], 2) - 2),
             '',
-            'Pending'
+            'Pending',
+            E.[TraceContext]
         FROM @NewOrchestrationEvents E
         WHERE NOT EXISTS (
             SELECT 1
             FROM Instances I
             WHERE [TaskHub] = @TaskHub AND I.[InstanceID] = E.[InstanceID])
-        GROUP BY E.[InstanceID]
+        GROUP BY E.[InstanceID], E.[TraceContext]
         ORDER BY E.[InstanceID] ASC
     END TRY
     BEGIN CATCH
@@ -1030,6 +1131,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     ) 
     SELECT 
@@ -1041,6 +1143,7 @@ BEGIN
         [RuntimeStatus],
         [VisibleTime],
         [TaskID],
+        [TraceContext],
         [PayloadID]
     FROM @NewOrchestrationEvents
 
@@ -1079,7 +1182,8 @@ BEGIN
         CASE WHEN @FetchOutput = 1 THEN (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = I.[InstanceID] AND
-            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText]
+            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText],
+        I.[TraceContext]
     FROM Instances I
     WHERE
         I.[TaskHub] = @TaskHub AND
@@ -1125,7 +1229,8 @@ BEGIN
         CASE WHEN @FetchOutput = 1 THEN (SELECT TOP 1 [Text] FROM Payloads P WHERE
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = I.[InstanceID] AND
-            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText]
+            P.[PayloadID] = I.[OutputPayloadID]) ELSE NULL END AS [OutputText],
+        I.[TraceContext]
     FROM
         Instances I
     WHERE
@@ -1189,7 +1294,8 @@ BEGIN
             P.[TaskHub] = @TaskHub AND
             P.[InstanceID] = N.[InstanceID] AND
             P.[PayloadID] = N.[PayloadID]) AS [PayloadText],
-        DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime]
+        DATEDIFF(SECOND, [Timestamp], @now) AS [WaitTime],
+        [TraceContext]
     FROM NewTasks N
     WHERE [TaskHub] = @TaskHub AND [SequenceNumber] = @SequenceNumber
 
@@ -1366,7 +1472,7 @@ BEGIN
     -- Instance IDs can be overwritten only if the orchestration is in a terminal state
     IF @existingStatus NOT IN ('Failed')
     BEGIN
-        DECLARE @msg nvarchar(4000) = FORMATMESSAGE('Cannot rewind instance with ID ''%s'' because it is not in a ''Failed'' state, but in ''%s'' state.', @InstanceID, @existingStatus);
+        DECLARE @msg nvarchar(4000) = FORMATMESSAGE('Cannot rewing instance with ID ''%s'' because it is not in a ''Failed'' state, but in ''%s'' state.', @InstanceID, @existingStatus);
         THROW 50001, @msg, 1;
     END